Support macro functions which do not need a semicolon at the end
Some macros declare the statement ending semicolon themselves. For example the
Qt library declares the Q_DISABLE_COPY macro like this:

    #define Q_DISABLE_COPY(Class) \
        Class(const Class &) Q_DECL_EQ_DELETE;\
        Class &operator=(const Class &) Q_DECL_EQ_DELETE;

When used like this:

    class Foo {
    private:
        Q_DISABLE_COPY(Foo)
        Bar* bar;
    };

uncrustify parser is confused and interprets the `*` as an arithmetic operator.

This patch adds a new option: `macro-func-nosm` which makes it possible to
declare some macros as not requiring an ending semicolon.